Move fonts instead of hardlinking them
