+++ lib/hbc/artifact/font.rb
@@ -1,4 +1,4 @@
-require "hbc/artifact/hardlinked"
+require "hbc/artifact/moved"
 
-class Hbc::Artifact::Font < Hbc::Artifact::Hardlinked
+class Hbc::Artifact::Font < Hbc::Artifact::Moved
 end
